Fix cicular depedency
