+++ src/ctapipe/irf/optimize/__init__.py
@@ -1,0 +1,15 @@
+from .algorithm import (
+    GhPercentileCutCalculator,
+    PercentileCuts,
+    PointSourceSensitivityOptimizer,
+    ThetaPercentileCutCalculator,
+)
+from .results import OptimizationResult
+
+__all__ = [
+    "OptimizationResult",
+    "GhPercentileCutCalculator",
+    "PercentileCuts",
+    "PointSourceSensitivityOptimizer",
+    "ThetaPercentileCutCalculator",
+]
